MAINT: interpolate: rename interpnd -> _interpnd
Also add an import shim with the deprecation message.
