+++ mypy.ini
@@ -122,7 +122,7 @@ ignore_missing_imports = True
 [mypy-scipy.interpolate._dfitpack]
 ignore_missing_imports = True
 
-[mypy-scipy.interpolate.interpnd]
+[mypy-scipy.interpolate._interpnd]
 ignore_missing_imports = True
 
 [mypy-scipy.interpolate._bspl]
